==> 01a_Habitat_Balancing.R <==
setwd("C:/Users/JordanGoetze/OneDrive - Department of Biodiversity, Conservation and Attractions/Research/Manuscripts/Manuscript_GlobalFinPrint Marine Reserves/Submission/Data")
dir()

detach("package:plyr", unload=TRUE)#will error - don't panic - need to make sure it is not loaded as will interfer with dplyr()
library(tidyr)
library(dplyr)
library(rstatix)
library(reshape2)
library(magrittr)

dat = read.csv("habitat_data.csv",header = T,strip.white = T)

#remove samples with unbalanced depth. Santo Domingo double zero so didn't need to balance
toremove <- scan(text="FP_2016_PH_01_MISAC_011 FP_2016_AU-I_01_ASAW_010 FP_2016_AU-I_01_ASAW_009 FP_2016_AU-I_01_ASAW_042 FP_2016_AU-I_01_ASAW_040
FP_2016_AU-I_01_ASAW_041 FP_2016_AU-I_01_ASAW_019 FP_2016_AU-I_01_ASAW_020 FP_2016_AU-I_01_ASAW_006 FP_2016_AU-I_01_ASAE_074
FP_2016_AU-I_01_ASAE_083 FP_2016_AU-I_01_ASAW_028 FP_2016_AU-I_01_ASAW_018 FP_2016_AU-I_01_ASAE_071
FP_2016_AU-I_01_ASAE_076 FP_2016_AU-I_01_ASAE_099 FP_2016_AU-I_01_ASAE_100 FP_2016_AU-I_01_ASAE_103
FP_2016_AU-I_01_ASAE_080 FP_2016_AU-I_01_ASAE_082 FP_2016_AU-I_01_ASAE_101 FP_2016_AU-I_01_ASAE_110
FP_2016_AU-I_01_ASAE_078 FP_2016_AU-I_01_ASAE_084 FP_2016_AU-I_01_ASAE_098 FP_2016_AU-I_01_ASAW_027
FP_2016_AU-I_01_ASAW_104 FP_2016_AU-I_01_ASAE_070 FP_2016_AU-I_01_ASAE_105 FP_2016_AU-I_01_ASAE_073
FP_2016_AU-I_01_ASAE_077 FP_2016_AU-I_01_ASAE_088 FP_2016_AU-I_01_ASAE_106 FP_2016_AU-I_01_ASAE_109
FP_2016_AU-I_01_ASAE_085 FP_2016_AU-I_01_ASAE_081 FP_2016_AU-I_01_ASAE_086 FP_2016_AU-I_01_ASAE_095
FP_2016_AU-I_01_ASAW_021 FP_2016_AU-I_01_ASAE_093 FP_2016_AU-I_01_ASAE_094 FP_2016_AU-I_01_ASAE_059
FP_2016_AU-I_01_ASAE_092 FP_2016_AU-I_01_ASAE_097 FP_2016_AU-I_01_ASAE_067 FP_2016_AU-I_01_ASAE_066
FP_2016_AU-I_01_ASAE_068 FP_2016_AU-I_01_ASAE_058 FP_2016_AU-I_01_ASAE_069 FP_2016_AU-I_01_ASAE_108
FP_2016_AU-I_01_ASAW_007 FP_2016_AU-I_01_ASAW_029 FP_2016_AU-I_01_ASAE_112 FP_2016_AU-I_01_ASAW_013 FP_2016_AU-I_01_ASAE_055
FP_2016_AU-I_01_ASAW_026 FP_2016_AU-I_01_ASAW_033 FP_2016_AU-I_01_ASAE_060 FP_2016_AU-I_01_ASAE_102
FP_2016_AU-I_01_ASAW_052 FP_2016_AU-I_01_ASAW_014 FP_2016_AU-I_01_ASAW_008 FP_2016_AU-I_01_ASAW_063
FP_2016_AU-I_01_ASAE_087 FP_2016_AU-I_01_ASAW_023 HE_2013_AU-P_01_LHBS_020 HE_2013_AU-P_01_LHBS_014
FP_2016_TT_04_TBBS_017 FP_2016_TT_04_TBBS_036 FP_2016_TT_04_TBBS_018 FP_2016_TT_04_TBBS_001
FP_2016_TT_04_TBBS_026 FP_2016_TT_04_TBBS_027 FP_2016_TT_04_TBBS_013 FP_2016_TT_04_TBBS_005
FP_2016_TT_04_TBBS_033 FP_2016_TT_04_TBBS_032 FP_2016_TT_04_TBBS_034 FP_2016_TT_04_TBBS_021
FP_2016_TT_04_TBBS_020 FP_2016_TT_04_TBBS_029 FP_2016_TT_04_TBBS_031 FP_2016_TT_04_TBBS_022
FP_2016_TT_04_TBBS_010 FP_2016_TT_04_TBBN_003 FP_2016_TT_04_TBBN_004 FP_2016_TT_04_TBBN_005
L_2015_DC_01_SXDF_006 L_2015_DC_01_SXDF_013 L_2015_DC_01_SXDF_029 L_2015_DC_01_SXDF_011
HE_2013_AU-P_01_LHLH_031 HE_2013_AU-P_01_LHLH_032 HE_2013_AU-P_01_LHLH_027 HE_2013_AU-P_01_LHLH_028
HE_2013_AU-P_01_LHLH_024 HE_2013_AU-P_01_LHLH_023 HE_2013_AU-P_01_LHLH_026 FP_2017_ID_03_GITO_046
FP_2017_ID_03_GITO_042 FP_2017_ID_03_GIMO_008 FP_2016_ZA_01_SL_iSN_036 FP_2016_ZA_01_SL_iSN_031
FP_2016_ZA_01_SL_iSN_027 FP_2016_ZA_01_SL_iSN_035 FP_2016_ZA_01_SL_iSN_116 FP_2016_ZA_01_SL_iSN_118
FP_2016_ZA_01_SL_iSN_026 FP_2016_ZA_01_SL_iSN_117 FP_2016_ZA_01_SL_iSN_039 FP_2016_ZA_01_SL_iSN_030
FP_2016_ZA_01_SL_iSN_025 FP_2016_ZA_01_SL_iSN_037 FP_2016_ZA_01_SL_iSN_034 FP_2016_ZA_01_SL_iSN_033
FP_2016_ZA_01_SL_iSN_029 FP_2016_ZA_01_SL_iSN_028 FP_2016_ZA_01_SL_iSN_032 FP_2016_ZA_01_SL_iSN_038
FP_2016_ZA_01_SL_iSN_040 FP_2016_ZA_01_SL_iSN_004 FP_2016_ZA_01_SL_iSN_081 FP_2016_ZA_01_SL_iSN_048
FP_2016_ZA_01_SL_iSN_021 FP_2016_ZA_01_SL_iSN_023 FP_2016_AU-P_04_LKHL_014 FP_2016_AU-P_04_LKHL_030
L_2012_DC_01_SBE_031 L_2012_DC_01_SBW_041 L_2012_DC_01_SBW_044 L_2012_DC_01_SBE_027
L_2012_DC_01_SBW_038 L_2012_DC_01_SBW_035 L_2012_DC_01_SBE_023 L_2012_DC_01_SBE_025
L_2012_DC_01_SBE_029 L_2012_DC_01_SBW_036 L_2012_DC_01_SBW_039 L_2012_DC_01_SBE_026
L_2012_DC_01_SBE_030 L_2012_DC_01_SBW_040 L_2012_DC_01_SBW_042 L_2012_DC_01_SBW_043
L_2012_DC_01_SBW_034 L_2012_DC_01_SBW_037 L_2012_DC_01_SBE_024 L_2012_DC_01_SBE_028
L_2015_DC_01_SXFF_030 L_2015_DC_01_SXFF_031 L_2015_DC_01_SXFF_032
FP_2017_JM_02_JMOR_023 FP_2017_JM_02_JMOR_027 FP_2017_JM_02_JMOR_021 FP_2017_JM_02_JMOR_001 FP_2017_JM_02_JMOR_022 FP_2017_JM_02_JMOR_006 FP_2017_JM_02_JMOR_025
FP_2017_JM_02_JMOR_009 FP_2017_JM_02_JMOR_024 FP_2017_JM_02_JMOR_005 FP_2017_JM_02_JMOR_020
FP_2017_JM_02_JMOR_007 FP_2017_BR_01_FNAP_066 FP_2017_BR_01_FNAP_022 FP_2017_BR_01_FNAP_067 FP_2017_BR_01_FNAP_068
L_2013_DC_01_SUR_034 L_2013_DC_01_SUR_023 L_2013_DC_01_SUR_036 L_2013_DC_01_SUR_035
L_2013_DC_01_SUR_038 L_2013_DC_01_SUR_043 L_2013_DC_01_SUR_021 L_2013_DC_01_SUR_026
L_2013_DC_01_SUR_028 L_2013_DC_01_SUR_031 L_2013_DC_01_SUR_039 L_2013_DC_01_SUR_044
L_2013_DC_01_SUR_016 L_2013_DC_01_SUR_002 L_2013_DC_01_SUR_004 L_2013_DC_01_SUR_015
L_2013_DC_01_SUR_001 L_2013_DC_01_SUR_010 L_2013_DC_01_SUR_012 L_2013_DC_01_SUR_014
L_2013_DC_01_SUR_003 L_2013_DC_01_SUR_005 L_2013_DC_01_SUR_011 L_2013_DC_01_SUE_023
L_2013_DC_01_SUE_025 L_2013_DC_01_SUE_026 L_2013_DC_01_SUE_027 L_2013_DC_01_SUE_030
L_2013_DC_01_SUE_033 L_2013_DC_01_SUE_035 L_2013_DC_01_SUE_029 L_2013_DC_01_SUE_032
L_2013_DC_01_SUE_028
", what="")

dat2 <- dat %>%
  dplyr::filter(!unique_id%in% toremove)%>%
  dplyr::collect()%>% #putting the collect at the ends keeps the query to the db as small as possible
  glimpse()

stat.test <- dat2 %>%
  group_by(unique_reserve) %>%
  t_test(depth ~ protection_status,paired = F,conf.level = 0.95,var.equal = F) %>%
  add_significance()
stat.test

stat.test.depth = stat.test

write.csv(stat.test.depth,"t-tests_depth.csv")

#Now check vis note will still be significant as only removing if <5m
toremovevis <- scan(text="FP_2016_PH_01_MISAO_004 LH_2015_AU-I_06_NCR_011
LH_2015_AU-I_06_NCR_009 LH_2015_AU-I_06_NCR_013 LH_2015_AU-I_06_NCR_008
LH_2015_AU-I_06_NCR_012 LH_2015_AU-I_06_NCR_003 FP_2016_CO_01_BTi_044
L_2015_DC_01_SXDF_039 LH_2015_AU-I_01_HAEO_014 LH_2015_AU-I_01_HAEO_012
LH_2015_AU-I_01_HAEO_011 LH_2015_AU-I_01_HAEO_015 LH_2015_AU-I_01_HAEO_006
LH_2015_AU-I_01_HAEO_008 LH_2015_AU-I_01_HAEO_003 LH_2015_AU-I_01_HAEO_009
LH_2015_AU-I_01_HAEO_020 LH_2015_AU-I_01_HAEO_030 LH_2015_AU-I_01_HAEO_021
FP_2017_US-P_01_MID_031 FP_2017_US-P_01_MID_026 FP_2017_US-P_01_MID_036
FP_2017_US-P_01_MID_034 FP_2017_US-P_01_MID_035 FP_2016_FJ_01_KBNM_007
FP_2016_FJ_02_KBNSR_038 FP_2016_FJ_02_KBNSR_036 FP_2016_FJ_02_KBNSR_035
LH_2015_AU-I_01_HAPR_002 LH_2015_AU-I_01_HAPO_014 LH_2015_AU-I_01_HAPO_015
LH_2015_AU-I_01_HAPO_021 LH_2015_AU-I_01_HAPO_020 LH_2015_AU-I_01_HAPO_011
LH_2015_AU-I_01_HAPO_003 LH_2015_AU-I_01_HAPO_012 LH_2015_AU-I_01_HAPO_006
LH_2015_AU-I_01_HAPO_027 FP_2017_LK_02_EPI_015 FP_2017_LK_02_EPI_021
FP_2015_MY_01_SSMK_067 FP_2015_MY_01_SSMK_070 FP_2015_MY_01_SSMK_065
L_2013_DC_01_SUR_035 L_2015_DC_01_SXFR_010 L_2015_DC_01_SXFR_016
L_2015_DC_01_SXFF_046 LH_2015_AU-I_01_HAWR_001 LH_2015_AU-I_01_HAWO_012
LH_2015_AU-I_01_HAWO_026 LH_2015_AU-I_01_HAWO_001 LH_2015_AU-I_01_HAWO_004
LH_2015_AU-I_01_HAWO_010 LH_2015_AU-I_01_HAWO_003 LH_2015_AU-I_01_HAWO_008
LH_2015_AU-I_01_HAWO_013 LH_2015_AU-I_01_HAWO_005 LH_2015_AU-I_01_HAWO_006
LH_2015_AU-I_01_HAWO_009 LH_2015_AU-I_01_HAWO_014 LH_2015_AU-I_01_HAWO_018
LH_2015_AU-I_01_HAWO_020 LH_2015_AU-I_01_HAWO_023 FP_2016_CO_01_RTe_033
FP_2016_CO_01_RGr_002 FP_2016_CO_01_RGr_030 FP_2016_CO_01_RGr_050
", what="")

dat3 <- dat2 %>%
  dplyr::filter(!unique_id%in% toremovevis)%>%
  dplyr::collect()%>% #putting the collect at the ends keeps the query to the db as small as possible
  glimpse()

# Remove NAs and reserves where vis is all the same (i.e. consistently good)
visdat = filter(dat3,!visibility == "NA")
visdat = subset(visdat, !unique_reserve %in% c("Balls Pyramid","Caye Caulker","Glover's Reef","MNP-23-1164","Namuri","Tanikely","Tetaiuo","Oracabessa"))
unique(visdat$unique_reserve)

#write.csv(visdat,"vis_data.csv")

stat.test <- visdat %>%
  group_by(unique_reserve) %>%
  t_test(visibility ~ protection_status,paired = F,conf.level = 0.95,var.equal = F) %>%
  add_significance()
stat.test

stat.test.vis = stat.test

#write.csv(stat.test.vis,"t-tests_vis.csv")

#Re-test depth
depthdat2 = subset(visdat, !unique_reserve %in% c("North Eustatius"))
stat.test <- depthdat2 %>%
  group_by(unique_reserve) %>%
  t_test(depth ~ protection_status,paired = F,conf.level = 0.95,var.equal = F) %>%
  add_significance()
stat.test

stat.test.depth = stat.test

#save lists of drops to remove in the data filtering script
write.csv(toremovevis,"vis_to_remove.csv")
write.csv(toremove,"depth_to_remove.csv")
